move component tests to components dir
