+++ src/__tests__/components/Button.test.js
@@ -1,6 +1,6 @@
 import { screen, render } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import Button from '../components/Button';
+import Button from '../../components/Button';
 
 const mockHandler = jest.fn(name => name);
